Restructure and additions to the keyboard driver + printf
Restructured everything to make more sense + work on the keyboard driver and added kprintf.
Also added neat splashscreen
