+++ include/types.h
@@ -1,0 +1,25 @@
+#pragma once
+
+typedef signed char     int8_t;
+typedef signed short    int16_t;
+typedef signed int      int32_t;
+typedef signed long     int64_t;
+
+typedef unsigned char   uint8_t;
+typedef unsigned short  uint16_t;
+typedef unsigned int    uint32_t;
+typedef unsigned long   uint64_t;
+
+typedef uint64_t        size_t;
+typedef uint16_t        wchar_t;
+typedef uint32_t        uintptr_t;
+
+typedef unsigned long int addr;
+
+typedef uint32_t        bool;
+
+#define false           0
+#define true            1
+
+#define NULL            0L
+#define NULL_POINTER    ((void*)0)
